add fix problem after Nurs Discuss

+++ src/modules/iban/auto.controller.ts
@@ -7,11 +7,14 @@ import {
   HttpStatus,
   Post,
   Query,
+  Req,
 } from '@nestjs/common';
 import { ApiResponse, ApiTags } from '@nestjs/swagger';
 
 import { RoleType } from '../../common/constants/role-type';
 import { Auth } from '../../decorators/http.decorators';
+import { NotValidCertException } from '../../exceptions/not-valid-cert.eception';
+import { SignatureService } from '../signature/signature.service';
 import { AutoService } from './auto.service';
 import { AutoDto } from './dto/auto.dto';
 import { AutoListDto } from './dto/autoList.dto';
@@ -19,7 +22,10 @@ import type { ISearchAccountResponse } from './interfaces/ISearchAccountResponse
 @Controller('auto')
 @ApiTags('auto')
 export class AutoController {
-  constructor(public readonly autoService: AutoService) {}
+  constructor(
+    public readonly autoService: AutoService,
+    public readonly signatureService: SignatureService,
+  ) {}
   @Post()
   @Auth(RoleType.USER)
   @HttpCode(HttpStatus.OK)
@@ -28,14 +34,45 @@ export class AutoController {
     description: 'Create auto in external service',
     type: AutoDto,
   })
-  addAutoAccount(@Body() autoDto: AutoDto) {
-    return this.autoService.addAutoAccount(autoDto);
+  async addAutoAccount(@Body() autoDto: AutoDto, @Req() req) {
+    const signatureData = await this.signatureService.verifySignature(
+      autoDto.signature,
+    );
+    if (
+      signatureData.valid &&
+      signatureData.subject.bin &&
+      signatureData.subject.bin === req.user.company.bin &&
+      signatureData.subject.iin === req.user.idn
+    ) {
+      return this.autoService.addAutoAccount(autoDto);
+    } else {
+      throw new NotValidCertException(
+        'Данные текущего пользователя и сертификата не совпадают',
+      );
+    }
   }
   @Delete()
   @Auth(RoleType.USER)
   @HttpCode(HttpStatus.OK)
-  deleteAutoAccount(@Body() autoDto: AutoDto): Promise<ISearchAccountResponse> {
-    return this.autoService.deleteAutoAccount(autoDto);
+  async deleteAutoAccount(
+    @Body() autoDto: AutoDto,
+    @Req() req: any,
+  ): Promise<ISearchAccountResponse> {
+    const signatureData = await this.signatureService.verifySignature(
+      autoDto.signature,
+    );
+    if (
+      signatureData.valid &&
+      signatureData.subject.bin &&
+      signatureData.subject.bin === req.user.company.bin &&
+      signatureData.subject.iin === req.user.idn
+    ) {
+      return this.autoService.deleteAutoAccount(autoDto);
+    } else {
+      throw new NotValidCertException(
+        'Данные текущего пользователя и сертификата не совпадают',
+      );
+    }
   }
   @Post('getList')
   @Auth(RoleType.USER)
